Changing the class structure around a smidge to make things easier
