+++ package.json
@@ -17,10 +17,10 @@
         "start": "ts-node-dev src/dev.ts",
         "lint": "eslint --ext .ts src",
         "lint:fix": "eslint --ext .ts src --fix",
-        "unpack": "ts-node-dev src/unpacker.ts",
-        "unpacker": "ts-node-dev src/unpacker.ts",
-        "index": "ts-node-dev src/indexer.ts",
-        "indexer": "ts-node-dev src/indexer.ts",
+        "unpack": "ts-node-dev --max-old-space-size=4096 src/unpacker.ts",
+        "unpacker": "ts-node-dev --max-old-space-size=4096 src/unpacker.ts",
+        "index": "ts-node-dev --max-old-space-size=4096 src/indexer.ts",
+        "indexer": "ts-node-dev --max-old-space-size=4096 src/indexer.ts",
         "copy-documents": "copyfiles package.json README.md .npmignore LICENSE lib",
         "package": "rimraf lib && npm i && npm run build && npm run copy-documents && cd lib && npm publish --dry-run",
         "publish:next": "npm run package && cd lib && npm publish -tag next",
