+++ src/fs/file-properties.ts
@@ -2,12 +2,10 @@ import { CompressionMethod } from '@runejs/common/compress';
 import { EncryptionMethod } from '@runejs/common/encrypt';
 import { Archive, FileIndex, Group, Store } from './index';
 import { setObjectProps } from '@runejs/common/util';
-import { IndexEntity } from '../db';
 
 
 export class FileProperties {
     key: string;
-    index: IndexEntity;
 
     store: Store | null = null;
     archive: Archive | null = null;
@@ -27,12 +25,8 @@ export class FileProperties {
     stripes: number[] = [];
     stripeCount: number = 1;
 
-    public constructor(properties?: Partial<FileProperties>) {
+    protected constructor(properties?: Partial<FileProperties>) {
         setObjectProps<FileProperties>(this, properties);
-
-        if(properties.index) {
-            this.index = properties.index;
-        }
     }
 
     public get numericKey(): number {
